refactor: split MobileFullscreenPlayer (831 lines) into focused subcomponents
Extract PlayerControls, PlayerProgress, PlayerTrackInfo, PlayerActions,
and PlayerLyricsContent. Main component delegates to subcomponents.
